move send transfer page component to send transfer side panel
