+++ src/domains/exchange/components/SendExchangeTransfer/SendExchangeTransfer.tsx
@@ -14,8 +14,8 @@ import { Modal } from "@/app/components/Modal";
 import { Networks } from "@/app/lib/mainsail";
 import { SelectAddress } from "@/domains/profile/components/SelectAddress";
 import { TotalAmountBox } from "@/domains/transaction/components/TotalAmountBox";
-import { TransferLedgerReview } from "@/domains/transaction/pages/SendTransfer/LedgerReview";
-import { buildTransferData } from "@/domains/transaction/pages/SendTransfer/SendTransfer.helpers";
+import { TransferLedgerReview } from "@/domains/transaction/components/SendTransferSidePanel/LedgerReview";
+import { buildTransferData } from "@/domains/transaction/components/SendTransferSidePanel/SendTransfer.helpers";
 import { calculateGasFee } from "@/domains/transaction/components/InputFee/InputFee";
 import { isLedgerTransportSupported } from "@/app/contexts/Ledger/transport";
 import { useLedgerContext } from "@/app/contexts";
